refactor(core): extract graph-recall coordinator (seam 14) (#1526)
Extract the graph-recall expansion subsystem from orchestrator.ts into a
dedicated GraphRecallCoordinator. Moves expandResultsViaGraph (spreading-
activation traversal from seed memories), recordLastGraphRecallSnapshot
(snapshot persistence), and two graph-seed-path helpers, plus 4 utility
functions (mergeGraphExpandedResults, graphPathRelativeToStorage,
blendGraphExpandedRecallScore, normalizeGraphActivationScore) and 2 type
interfaces (GraphRecallRankedResult, GraphRecallShadowComparison).

Behavior-preserving: the orchestrator constructs one coordinator instance
with getter/callback accessors (matching the RecallRerankCoordinator /
ContradictionLinkingCoordinator pattern) and keeps thin delegating methods
so all call sites (recallInternal, cold-fallback pipeline, tests) work
unchanged.

orchestrator.ts: 18518 → 18200 LOC (−318)
